// src/redux/game/game.types.ts
enum GameTypes {
	RESET_GAME = 'RESET_GAME',
	SET_UP_GAME_START = 'SET_UP_GAME_START',
	SET_UP_GAME_SUCCESS = 'SET_UP_GAME_SUCCESS',
	SELECT_PLOT_START = 'SELECT_PLOT_START',
	SELECT_PLOT_SUCCESS = 'SELECT_PLOT_SUCCESS',
	MARK_PLOT_START = 'MARK_PLOT_START',
	MARK_PLOT_SUCCESS = 'MARK_PLOT_SUCCESS',
	GAME_OVER = 'GAME_OVER',
	UPDATE_STATS = 'UPDATE_STATS',
}

export default GameTypes;